LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/navigate-204-205-download-expected.txt


FAIL navigate() promises to 204s never settle assert_equals: expected 1 but got 0
FAIL navigate() promises to 205s never settle assert_equals: expected 1 but got 0
FAIL navigate() promises to Content-Disposition: attachment responses never settle assert_equals: expected 1 but got 0
